1. Add addCar button to cars page 2. Add create element logic to commons js 3. Fix date and time problems in car save 4. Add consts object to $.hib
Former-commit-id: 6c3b8ac9b8cbbebc1b44452a42260fd4d4a295f3

+++ public/javascripts/clientCommon.js
@@ -1,5 +1,12 @@
 $.hib = {};
 
+$.hib.consts = {
+	pickATimeDateFormat: "dd-mm-yyyy",
+	pickATimeTimeFormat: "H:i",
+	momentDateFormat: "DD-MM-YYYY",
+	momentTimeFormat: "H:m"
+};
+
 $.hib.post = function(url, data, success, error) {
 	$.ajax({
 		type: "POST",
@@ -35,16 +42,23 @@ $.hib.put = function(url, data, success, error) {
 		3. Has a td that has the three icons for delete, edit and approve changes
 
 	The two last parameters are functions for additional actions to run on rowEditable or rowNonEditable events
- */
-$.hib.makeTableRowsEditable = function(modelName, buildObjectToUpdateFunc, makeRowEditableAdditionalActions, makeRowNonEditableAdditionalActions) {
-	$('.finishedEditingIcon').hide();
+	*/
+	$.hib.makeTableRowsEditable = function(modelName, buildObjectToUpdateFunc, makeRowEditableAdditionalActions, makeRowNonEditableAdditionalActions) {
+		$('.finishedEditingIcon').hide();
 
-	$('.deleteIcon').off('click').on('click', function() {
-		var parentRowElement = $(this).closest('tr');
-		bootbox.confirm('האם למחוק?', function(userWantsToDelete) {
-			if (userWantsToDelete) {
-				var objectToDeleteId = parentRowElement.attr('id');
-				$.ajax({
+		$('.deleteIcon').off('click').on('click', function() {
+			var parentRowElement = $(this).closest('tr');
+			bootbox.confirm('האם למחוק?', function(userWantsToDelete) {
+				if (userWantsToDelete) {
+					var objectToDeleteId = parentRowElement.attr('id');
+
+					// If this element is not yet saved in the DB, just remove the row in UI.
+					if (!objectToDeleteId) {
+						parentRowElement.remove();
+						return;
+					}
+
+					$.ajax({
 					// e.g /consumers/72
 					url: '/' + modelName + 's/' + objectToDeleteId,
 					type: 'DELETE',
@@ -56,55 +70,76 @@ $.hib.makeTableRowsEditable = function(modelName, buildObjectToUpdateFunc, makeR
 						bootbox.alert("Failed deleting " + modelName);
 					}
 				});
+				}
+			});
+		});
+
+		$('.editIcon').off('click').on('click', function() {
+			var editIconElement = $(this);
+			var deleteIconElement = editIconElement.siblings('.deleteIcon');
+			var finishedEditingIconElement = editIconElement.siblings('.finishedEditingIcon');
+
+			editIconElement.hide();
+			deleteIconElement.hide();
+			finishedEditingIconElement.show();
+
+			var parentRowElement = editIconElement.closest('tr');
+			_makeRowEditable(parentRowElement);
+
+		});
+
+		$('.finishedEditingIcon').off('click').on('click', function() {
+			var finishedEditingIconElement = $(this);
+			var deleteIconElement = finishedEditingIconElement.siblings('.deleteIcon');
+			var editIconElement = finishedEditingIconElement.siblings('.editIcon');
+
+			finishedEditingIconElement.hide();
+			deleteIconElement.show();
+			editIconElement.show();
+
+			var parentRowElement = finishedEditingIconElement.closest('tr');
+			_makeRowNotEditable(parentRowElement);
+
+			// Update server
+			if (buildObjectToUpdateFunc) {
+				var objectToUpdate = buildObjectToUpdateFunc(parentRowElement);
+				if (!objectToUpdate) {
+					return;
+				}
+
+				var objectToUpdateId = parentRowElement.attr('id');
+				var url = '/' + modelName + 's';
+				var success, error;
+				
+				// If object has id update, otherwise create
+				if (objectToUpdateId) {
+					url = url + '/' + objectToUpdateId;
+
+					success = function() {
+						console.log("Updated " + modelName + " with id " + objectToUpdateId);
+						parentRowElement.removeClass('italicText');
+					};
+					error = function() {
+						alert('failed updating ' + modelName + " with id " + objectToUpdateId);
+					};
+
+					parentRowElement.addClass('.italicText');
+					$.hib.put(url, objectToUpdate, success, error);	
+				} else {
+					success = function(createdElement) {
+						console.log("Created " + modelName);
+						parentRowElement.removeClass('italicText');
+						parentRowElement.attr('id', createdElement._id);
+					};
+					error = function() {
+						alert('Failed creating ' + modelName);
+					};
+
+					parentRowElement.addClass('.italicText');
+					$.hib.post(url, objectToUpdate, success, error);	
+				}
 			}
 		});
-	});
-
-	$('.editIcon').off('click').on('click', function() {
-		var editIconElement = $(this);
-		var deleteIconElement = editIconElement.siblings('.deleteIcon');
-		var finishedEditingIconElement = editIconElement.siblings('.finishedEditingIcon');
-
-		editIconElement.hide();
-		deleteIconElement.hide();
-		finishedEditingIconElement.show();
-
-		var parentRowElement = editIconElement.closest('tr');
-		_makeRowEditable(parentRowElement);
-
-	});
-
-	$('.finishedEditingIcon').off('click').on('click', function() {
-		var finishedEditingIconElement = $(this);
-		var deleteIconElement = finishedEditingIconElement.siblings('.deleteIcon');
-		var editIconElement = finishedEditingIconElement.siblings('.editIcon');
-
-		finishedEditingIconElement.hide();
-		deleteIconElement.show();
-		editIconElement.show();
-
-		var parentRowElement = finishedEditingIconElement.closest('tr');
-		_makeRowNotEditable(parentRowElement);
-
-		// Update server
-		if (buildObjectToUpdateFunc) {
-			var objectToUpdateId = parentRowElement.attr('id');
-			var url = '/' + modelName + 's/' + objectToUpdateId;
-			var objectToUpdate = buildObjectToUpdateFunc(parentRowElement);
-
-			var success = function() {
-				console.log("Updated " + modelName + " with id " + objectToUpdateId);
-				parentRowElement.removeClass('italicText');
-			};
-			var error = function() {
-				alert('failed updating ' + modelName + " with id " + objectToUpdateId);
-			};
-
-			parentRowElement.addClass('.italicText');
-			$.hib.put(url, objectToUpdate, success, error);
-		}
-
-	});
 
 	// Double click on row will be like clicking the edit icon of that row
 	$('.nonHeaderTableRow').off('click').on('dblclick', function(event) {
@@ -136,13 +171,20 @@ $.hib.makeTableRowsEditable = function(modelName, buildObjectToUpdateFunc, makeR
 
 	// When clicking outside a row we want to make it not editable
 	function _detectClickOutsideRowAndMakeItNotEditable(rowElement) {
+		var randomTempId = _generateRandomTempId();
+		rowElement.attr('tempId', randomTempId);
+		
 		$(document).off('click').on('click', function(event) { 
-			var rowId = rowElement.attr('id');
 			// If event target is not a child of the row, make row uneditable
-			if(!$(event.target).closest('#' + rowId).length) {
+			if(!$(event.target).closest('tr[tempId=' + randomTempId + ']').length) {
 				rowElement.find('.finishedEditingIcon').click();
 				$(document).unbind('click');
+				rowElement.removeAttr('tempId');
 			}      
 		});
+	}
+
+	function _generateRandomTempId() {
+		return Math.random() * 100000000;
 	}
 };
